% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizzBuzz}
\alias{fizzBuzz}
\title{A function that takes in a vector of numbers and applies the
FizzBuzz logic to each element. If a number is divisible by 3,
it is replaced with "Fizz". If it is divisible by 5,
it is replaced with "Buzz". If it is divisible by both 3 and 5,
it is replaced with "Fizz Buzz". Otherwise, the number remains the same.}
\usage{
fizzBuzz(nums)
}
\arguments{
\item{nums}{A vector of numbers to apply FizzBuzz to.}
}
\value{
A character vector representing the FizzBuzz output
}
\description{
A function that takes in a vector of numbers and applies the
FizzBuzz logic to each element. If a number is divisible by 3,
it is replaced with "Fizz". If it is divisible by 5,
it is replaced with "Buzz". If it is divisible by both 3 and 5,
it is replaced with "Fizz Buzz". Otherwise, the number remains the same.
}
