% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{data set of weather information at Flagstaff's Pulliam Airport from 1950 to 2019.}
\format{
a data frame with 22967 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{RAIN}{Daily percipitation.}
  \item{SNOW}{Daily snow.}
  \item{TMAX}{Daily max temp.}
  \item{TMIN}{Daily min temp.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
data set of weather information at Flagstaff's Pulliam Airport from 1950 to 2019.
}
\keyword{datasets}
